Direct use of io::Result for cleaner look & tests
