+++ steel-protocol/src/packets/shared_implementation.rs
@@ -1,4 +1,4 @@
-use std::io::{self, Write};
+use std::io::{self, Write, Result};
 
 use steel_utils::text::TextComponentBase;
 use uuid::Uuid;
@@ -6,14 +6,14 @@ use uuid::Uuid;
 use crate::packet_traits::{ReadFrom, WriteTo};
 
 impl WriteTo for TextComponentBase {
-    fn write(&self, _: &mut impl Write) -> Result<(), io::Error> {
+    fn write(&self, _: &mut impl Write) -> Result<()> {
         //TODO: Implement
         todo!()
     }
 }
 
 impl ReadFrom for Uuid {
-    fn read(data: &mut impl io::Read) -> Result<Self, io::Error> {
+    fn read(data: &mut impl io::Read) -> Result<Self> {
         let most_significant_bits = u64::read(data)?;
         let least_significant_bits = u64::read(data)?;
 
@@ -25,7 +25,7 @@ impl ReadFrom for Uuid {
 }
 
 impl WriteTo for Uuid {
-    fn write(&self, writer: &mut impl Write) -> Result<(), io::Error> {
+    fn write(&self, writer: &mut impl Write) -> Result<()> {
         let (most_significant_bits, least_significant_bits) = self.as_u64_pair();
         most_significant_bits.write(writer)?;
         least_significant_bits.write(writer)?;
