+++ steel-protocol/src/ser/prefixed_read.rs
@@ -1,4 +1,4 @@
-use std::io::{Error, Read};
+use std::io::{Error, Read, Result};
 
 use crate::packet_traits::{PrefixedRead, ReadFrom};
 
@@ -6,13 +6,13 @@ impl PrefixedRead for String {
     fn read_prefixed_bound<P: TryInto<usize> + ReadFrom>(
         data: &mut impl Read,
         bound: usize,
-    ) -> Result<Self, Error> {
+    ) -> Result<Self> {
         let len: usize = P::read(data)?
             .try_into()
             .map_err(|_| Error::other("Invalid Prefix"))?;
 
         if len > bound {
-            return Result::Err(Error::other("To long"));
+            Err(Error::other("To long"))?
         }
 
         let mut buf = vec![0; len];
@@ -25,13 +25,13 @@ impl<T: ReadFrom> PrefixedRead for Vec<T> {
     fn read_prefixed_bound<P: TryInto<usize> + ReadFrom>(
         data: &mut impl Read,
         bound: usize,
-    ) -> Result<Self, Error> {
+    ) -> Result<Self> {
         let len: usize = P::read(data)?
             .try_into()
             .map_err(|_| Error::other("Invalid Prefix"))?;
 
         if len > bound {
-            return Result::Err(Error::other("To long"));
+            Err(Error::other("To long"))?
         }
         let mut items = Vec::with_capacity(len);
         for _ in 0..len {
