+++ steel-protocol/src/ser/read.rs
@@ -1,16 +1,16 @@
-use std::io::{Error, Read};
+use std::io::{Read, Result};
 
 use crate::packet_traits::ReadFrom;
 
 impl ReadFrom for bool {
-    fn read(data: &mut impl Read) -> Result<Self, Error> {
+    fn read(data: &mut impl Read) -> Result<Self> {
         let byte = u8::read(data)?;
         Ok(byte == 1)
     }
 }
 
 impl ReadFrom for u8 {
-    fn read(data: &mut impl Read) -> Result<Self, Error> {
+    fn read(data: &mut impl Read) -> Result<Self> {
         let mut buf = [0; size_of::<Self>()];
         data.read_exact(&mut buf)?;
         Ok(Self::from_be_bytes(buf))
@@ -18,7 +18,7 @@ impl ReadFrom for u8 {
 }
 
 impl ReadFrom for u16 {
-    fn read(data: &mut impl Read) -> Result<Self, Error> {
+    fn read(data: &mut impl Read) -> Result<Self> {
         let mut buf = [0; size_of::<Self>()];
         data.read_exact(&mut buf)?;
         Ok(Self::from_be_bytes(buf))
@@ -26,7 +26,7 @@ impl ReadFrom for u16 {
 }
 
 impl ReadFrom for u32 {
-    fn read(data: &mut impl Read) -> Result<Self, Error> {
+    fn read(data: &mut impl Read) -> Result<Self> {
         let mut buf = [0; size_of::<Self>()];
         data.read_exact(&mut buf)?;
         Ok(Self::from_be_bytes(buf))
@@ -34,7 +34,7 @@ impl ReadFrom for u32 {
 }
 
 impl ReadFrom for u64 {
-    fn read(data: &mut impl Read) -> Result<Self, Error> {
+    fn read(data: &mut impl Read) -> Result<Self> {
         let mut buf = [0; size_of::<Self>()];
         data.read_exact(&mut buf)?;
         Ok(Self::from_be_bytes(buf))
@@ -42,7 +42,7 @@ impl ReadFrom for u64 {
 }
 
 impl ReadFrom for i8 {
-    fn read(data: &mut impl Read) -> Result<Self, Error> {
+    fn read(data: &mut impl Read) -> Result<Self> {
         let mut buf = [0; size_of::<Self>()];
         data.read_exact(&mut buf)?;
         Ok(Self::from_be_bytes(buf))
@@ -50,7 +50,7 @@ impl ReadFrom for i8 {
 }
 
 impl ReadFrom for i16 {
-    fn read(data: &mut impl Read) -> Result<Self, Error> {
+    fn read(data: &mut impl Read) -> Result<Self> {
         let mut buf = [0; size_of::<Self>()];
         data.read_exact(&mut buf)?;
         Ok(Self::from_be_bytes(buf))
@@ -58,7 +58,7 @@ impl ReadFrom for i16 {
 }
 
 impl ReadFrom for i32 {
-    fn read(data: &mut impl Read) -> Result<Self, Error> {
+    fn read(data: &mut impl Read) -> Result<Self> {
         let mut buf = [0; size_of::<Self>()];
         data.read_exact(&mut buf)?;
         Ok(Self::from_be_bytes(buf))
@@ -66,7 +66,7 @@ impl ReadFrom for i32 {
 }
 
 impl ReadFrom for i64 {
-    fn read(data: &mut impl Read) -> Result<Self, Error> {
+    fn read(data: &mut impl Read) -> Result<Self> {
         let mut buf = [0; size_of::<Self>()];
         data.read_exact(&mut buf)?;
         Ok(Self::from_be_bytes(buf))
@@ -74,7 +74,7 @@ impl ReadFrom for i64 {
 }
 
 impl ReadFrom for f32 {
-    fn read(data: &mut impl Read) -> Result<Self, Error> {
+    fn read(data: &mut impl Read) -> Result<Self> {
         let mut buf = [0; size_of::<Self>()];
         data.read_exact(&mut buf)?;
         Ok(Self::from_be_bytes(buf))
@@ -82,7 +82,7 @@ impl ReadFrom for f32 {
 }
 
 impl ReadFrom for f64 {
-    fn read(data: &mut impl Read) -> Result<Self, Error> {
+    fn read(data: &mut impl Read) -> Result<Self> {
         let mut buf = [0; size_of::<Self>()];
         data.read_exact(&mut buf)?;
         Ok(Self::from_be_bytes(buf))
